Move src/api to include/yajl.
This allows the source to be used as-is, without install step,
by adding a include path to yajl's include directory.
