+++ include/yajl/yajl_common.h
