+++ include/yajl/yajl_version.h.cmake
