+++ src/yajl_bytestack.h
@@ -22,7 +22,7 @@
 #ifndef __YAJL_BYTESTACK_H__
 #define __YAJL_BYTESTACK_H__
 
-#include "api/yajl_common.h"
+#include "yajl/yajl_common.h"
 
 #define YAJL_BS_INC 128
