+++ src/yajl_gen.c
@@ -14,7 +14,7 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-#include "api/yajl_gen.h"
+#include "yajl/yajl_gen.h"
 #include "yajl_buf.h"
 #include "yajl_encode.h"
